requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-openai
langchain-google-genai
langchain-groq
langchain-mistralai
python-dotenv
faiss-cpu
tiktoken
fastapi
uvicorn
requests
langchain_huggingface
pypdf
bs4